[Native]Remove redundant loops and unnecessary task launches
In `TaskManager.maybeStartNextQueuedTask`, if we found that one task
is gone or aborted, then we can directly skip the whole queue of tasks
that belongs to the same query; there is no need to continue the inner
loop and check the rest ones.

Besides, this change properly handles the clearing of `tasksToStart`,
preventing it from launching tasks that should not be started in
certain edge cases—for example, if there's only one batch and it has
an invalid task.